Move iterator tests to `clist_iterator_tests.go`
